Move broker-cli to its own top-level dir
This is in anticipation of releasing the broker-cli tool.
