+++ broker-cli/pkg/client/adapter/adapter.go
