+++ broker-cli/pkg/client/adapter/httpadapter_test.go
